Casting size to (int) to prevent warning
git-svn-id: http://svn/svn/hcr/trunk@498 2e3f476d-dfe7-4101-bf40-6ec35b1080dd

+++ src/hcrarchiver/CfRadialFilesReader.cpp
@@ -26,7 +26,7 @@ bool
 CfRadialFilesReader::getNextRay(const RadxRay* & radxRay, 
         const RadxVol* & radxVol) {
     // Open the next file if necessary
-    if (_nextRayNum == _radxVol.getNRays() && ! _openNextFile())
+  if (_nextRayNum == (int) _radxVol.getNRays() && ! _openNextFile())
             return false;
     
     radxRay = _radxVol.getRays()[_nextRayNum];
